+++ app/src/main/java/com/example/wateringreminder/compose/navigation/BottomNavGraph.kt
@@ -5,6 +5,7 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import com.example.wateringreminder.articlelist.MyPlantsScreen
+import com.example.wateringreminder.compose.navigation.BottomBarScreen
 import com.example.wateringreminder.watering.WateringScreen
